move file & star
