+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -45,7 +45,7 @@
 
     <fragment
         android:id="@+id/navigation_preference"
-        android:name="com.bignerdranch.android.popuptrip.ui.setting.PreferenceFragment"
+        android:name="com.bignerdranch.android.popuptrip.ui.Profile.PreferenceFragment"
         android:label="@string/title_pref"
         tools:layout="@layout/fragment_preference"/>
